correct text generator to handle missing of all the variables

+++ R/htk_text_generator.r
@@ -36,30 +36,42 @@ htk_text_generator=function(data,category,ranking,ctry,var_codes){
   main_vars_label<<-vars_needed$var_names_long
   main_vars_direction <<- vars_needed$var_direction
 
-  #select the variables in the database
-  mydt<<-vars_needed$data %>% data.frame()
-  mydt<<-mydt %>% dplyr::select(Iso_code3,main_vars,main_vars_rank)
+  if(length(main_vars)==3){
+    #select the variables in the database
+    mydt<<-vars_needed$data %>% data.frame()
+    mydt<<-mydt %>% dplyr::select(Iso_code3,main_vars,main_vars_rank)
 
-  output <-"paragraphs.R"
+    output <-"paragraphs.R"
 
-   current.folder=system.file("extdata", package = "OECDHousingToolkit")
-   local.folder <- "."
-   file.copy(paste0(current.folder,"/",output), local.folder,overwrite = T)
-  source(output)
+    current.folder=system.file("extdata", package = "OECDHousingToolkit")
+    local.folder <- "."
+    file.copy(paste0(current.folder,"/",output), local.folder,overwrite = T)
+    source(output)
 
-  #include the paragraphs in the database for each country
-  mydt=mydt %>% filter(Iso_code3==ctry)
-  mydt[[category]]=NA
-  for(i in 1:length(mydt$Iso_code3)){
-    i<<-i
-    ctry=mydt[i,"Iso_code3"]
-    current_input <- c(round(mydt[i,colsnum[1]],2),round(mydt[i,colsnum[2]],2),round(mydt[i,colsnum[3]],2))
-    my_ldcp <- ldcp_run(ldcp=my_ldcp,input=current_input)
-    text=gsub("targetcountry",countrycode(ctry,origin="iso3c",destination="country.name"),my_ldcp$report$description)
-    mydt[i,category]=text
+    #include the paragraphs in the database for each country
+    mydt=mydt %>% filter(Iso_code3==ctry)
+    mydt[[category]]=NA
+    for(i in 1:length(mydt$Iso_code3)){
+      i<<-i
+      ctry=mydt[i,"Iso_code3"]
+      current_input <- c(round(mydt[i,colsnum[1]],2),round(mydt[i,colsnum[2]],2),round(mydt[i,colsnum[3]],2))
+      my_ldcp <- ldcp_run(ldcp=my_ldcp,input=current_input)
+      text=gsub("targetcountry",countrycode(ctry,origin="iso3c",destination="country.name"),my_ldcp$report$description)
+      mydt[i,category]=text
+    }
+    #colnames(mydt)=c("country",main_vars,category)
+    file.remove(output)
+    #rio::export(mydt,paste0("htk_paragraphs_",category,".csv"),sep=";")
+  }else{
+    mydt=data.frame(ctry,
+                    paste0("The ",category, "of the housing sector in ",countrycode(ctry,origin="iso3c",destination="country.name")," cannot be assess based on available data.
+                           Improvments on data collection should be undertaken"),
+                    paste0("The ",category, "of the housing sector in ",countrycode(ctry,origin="iso3c",destination="country.name")," cannot be assess based on available data.
+                           Improvments on data collection should be undertaken"),
+                    paste0("The ",category, "of the housing sector in ",countrycode(ctry,origin="iso3c",destination="country.name")," cannot be assess based on available data.
+                           Improvments on data collection should be undertaken"))
+    colnames(mydt)=c("Iso_code3","efficiency","inclusiveness","sustainability")
   }
-  #colnames(mydt)=c("country",main_vars,category)
-   file.remove(output)
-  #rio::export(mydt,paste0("htk_paragraphs_",category,".csv"),sep=";")
+
   mydt
 }
